Add WhatsApp auth and query services
